requirements.txt
# requirements.txt

# Core
python-dotenv>=1.0.0

# Audio processing
pydub>=0.25.1

# Transcription
openai-whisper>=20231117

# LLM Client
requests>=2.31.0

# Document processing
python-docx>=1.1.0

# Flowchart generation
graphviz>=0.20.1

# Image/Video processing
opencv-python>=4.8.0
Pillow>=10.1.0

# Frontend
streamlit>=1.30.0

# Utilities
tqdm>=4.66.0